// strategy/common/ingest.d.ts
/* this code was generated by automated tool,
   should not edit by hand */

export interface IngestService {
  ingest(request: IngestRequest): Promise<IngestResponse>;
  ingest_arrow(request: IngestArrowRequest): Promise<IngestResponse>;
}

export type OperationType = "INSERT" | "DELETE" | "UPDATE" | string;

export interface IngestRequest {
  schemaName?: string;
  typ?: OperationType;
  old?: Value[];
  new?: Value[];
  seqNo?: number;
}

export interface IngestResponse {
  seqNo?: number;
}

export interface PointType {
  x?: number;
  y?: number;
}

export interface DurationType {
  value?: string;
  timeUnit?: string;
}

export interface RustDecimal {
  scale?: number;
  lo?: number;
  mid?: number;
  hi?: number;
  negative?: boolean;
}

export interface Timestamp {
  seconds?: number;
  nanos?: number;
}

export interface Value {
  uintValue?: number;
  uint_128Value?: string;
  intValue?: number;
  int_128Value?: string;
  floatValue?: number;
  boolValue?: boolean;
  stringValue?: string;
  bytesValue?: Uint8Array;
  decimalValue?: RustDecimal;
  timestampValue?: Timestamp;
  dateValue?: string;
  pointValue?: PointType;
  durationValue?: DurationType;
  jsonValue?: JsonValue;
}

export interface IngestArrowRequest {
  schemaName?: string;
  records?: Uint8Array;
  seqNo?: number;
  metadata?: IngestMetadata;
}

export interface IngestMetadata {
  typ?: OperationType;
  version?: number;
}

export interface Struct {
  fields?: JsonValue;
}

export interface JsonValue {
  nullValue?: NullValue;
  numberValue?: number;
  stringValue?: string;
  boolValue?: boolean;
  structValue?: Struct;
  listValue?: ListValue;
}

export type NullValue = "NULL_VALUE" | string;

export interface ListValue {
  values?: JsonValue[];
}